fix: fix all inconsistency in search
