+++ packages/web/utils/GlobalSearch.ts
@@ -2,4 +2,4 @@ export const GlobalFilters = {
   PLAYERS: 'Players',
   PATRONS: 'Patrons',
   GUILDS: 'Guilds',
-};
+} as const;
